MAINT: Moved Config object to config module
